lambda/hello_world/requirements.txt
requests
google-cloud-vision==2.1.0
azure-cosmos==4.2.0
boto3==1.16.62